tests: Added unit test for checklist item member use-cases

+++ api/src/application/use-cases/checklist-item-member/DeleteChecklistItemUseCase.js
@@ -4,10 +4,21 @@ class DeleteChecklistItemUseCase {
   }
 
   async execute(checklistItemId, projectMemberId) {
-    return this.checklistItemMemberRepository.delete(
+    if (!checklistItemId) throw new Error('checklistItemId was not provided');
+    if (!projectMemberId) throw new Error('projectMemberId was not provided');
+
+    const result = await this.checklistItemMemberRepository.delete(
       checklistItemId,
       projectMemberId,
     );
+
+    if (result === 0) {
+      throw new Error(
+        'Something went wrong removing the checklist item member',
+      );
+    }
+
+    return result;
   }
 }
